fixed declaration of GLFWcursorposfun. x and y used to be ints. They are now doubles.

+++ import/derelict/glfw3/types.d
@@ -300,7 +300,7 @@ extern(C)
     alias void function(GLFWwindow*, int) GLFWwindowfocusfun;
     alias void function(GLFWwindow*, int) GLFWwindowiconifyfun;
     alias void function(GLFWwindow*, int, int) GLFWmousebuttonfun;
-    alias void function(GLFWwindow*, int, int) GLFWcursorposfun;
+    alias void function(GLFWwindow*, double, double) GLFWcursorposfun;
     alias void function(GLFWwindow*, int) GLFWcursorenterfun;
     alias void function(GLFWwindow*, double, double) GLFWscrollfun;
     alias void function(GLFWwindow*, int, int) GLFWkeyfun;
